Update Space_Between.md
Added a little bit about Hawthorne and her presence in the Space Between. It seems like the Space Between is a step ahead of her this time--if she were to communicate with Ariadne even briefly, she would cause a memetic outbreak in the PRISM and undo everything I worked on in the revamp.